dictionaries placed in text files
