+++ transactions_at_places.txt
@@ -1,0 +1,3 @@
+"amazon": 0, "clothes": 0, "gas": 0, "cats": 0, "costco": 0, "grocery": 0, "internet": 0, "ups": 0, 
+"usps": 0, "utilities": 0, "apple": 0, "entertainment": 0, "food_delivery": 0, "target": 0, "office_supplies": 0, 
+"gym": 0,  "eating_out": 0, "other": 0, "allstate": 0
